+++ CHANGELOG.md
@@ -17,6 +17,7 @@ You can also check the
     others too – and we fall back to a regular string in case the label is not
     localized
 - Fixes
+  - Color palette can be changed again
   - Combo charts tooltips now work correctly again
 
 # [5.7.0] - 2025-04-16